DRAFT: split button.tsx into button/ folder
