+++ gatsby/src/components/button/A.tsx
@@ -1,0 +1,9 @@
+export default function A(props: ButtonProps & StyleProps) {
+  const className: string = styles.a + " " + props.className
+  return (
+    <a className={className} href={props.href}>
+      {props.text}
+      {props.children}
+    </a>
+  )
+}
