+++ gatsby/src/components/button/index.ts
@@ -1,0 +1,5 @@
+import A from './A.tsx'
+import NavAnchor from './NavAnchor.tsx'
+import ButtonArrow from './ButtonArrow.tsx'
+
+export { A, NavAnchor, ButtonArrow }
