+++ gatsby/src/components/button/types.ts
@@ -1,0 +1,5 @@
+export interface ButtonProps {
+  href?: string
+  text: string
+  right?: boolean
+}
