+++ gatsby/src/components/icon/SvgRightArrow.tsx
@@ -1,12 +1,9 @@
 import * as React from "react"
-import Rightarrow from "../images/svg/ra.svg"
+import Rightarrow from "../../images/svg/ra.svg"
+import { StyleProps } from "../../utils"
 import * as styles from "../styles/utilities/other.module.css"
 
-interface StyleProps {
-  className: string
-}
-
-export function SvgRightArrow(props: StyleProps) {
+export default function SvgRightArrow(props: StyleProps) {
   return (
     <span className={props.className}>
       <Rightarrow width={12} height={12} />
