+++ gatsby/src/components/icon/index.ts
@@ -1,0 +1,3 @@
+import SvgRightArrow from './SvgRightArrow.tsx'
+
+export { SvgRightArrow }
