+++ gatsby/src/types.ts
@@ -1,0 +1,9 @@
+export interface DataCyProps {
+  dataCy: string
+}
+
+export interface SvgProps {
+  icon: React.FunctionComponent<React.SVGProps<SVGSVGElement>>
+  title: string
+  ariaLabel: string
+}
